+++ packages/graph-explorer/src/core/ThemeProvider/index.ts
@@ -1,6 +1,4 @@
-export { default as ThemeProvider } from "./ThemeProvider";
-export { default as useTheme } from "./useTheme";
-export { default as useWithTheme } from "./useWithTheme";
+export * from "./ThemeProvider";
 
 // Color utils
 export { default as contrastColor } from "./utils/contrastColor";
@@ -8,18 +6,3 @@ export { default as darken } from "./utils/darken";
 export { default as fade } from "./utils/fade";
 export { default as lighten } from "./utils/lighten";
 export { default as saturate } from "./utils/saturate";
-
-export type {
-  ActiveThemeType,
-  Background,
-  Grey,
-  Palette,
-  ProcessedTheme,
-  Shape,
-  Spacing,
-  TextPalette,
-  ThemeContextType,
-  ThemeStyleFn,
-  ThemedStyle,
-  Typography,
-} from "./types";
